+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Added 'Power Saving' setting which turns off the AP and lowers the CPU frequency on Cardputer and S3TWatch.
   - Terminal App to use commands with the keyboard - @tototo31
   - Added option to select Custom Evil Portal html file from the SD Card - @tototo31
+  - Placeholder text for keyboard view - @tototo31
 
 - Commands
   - 'chipinfo' command to display chip information